TST: collect scalar Interval tests
