+++ pandas/tests/scalar/interval/test_overlaps.py
@@ -1,4 +1,3 @@
-"""Tests for Interval-Interval operations, such as overlaps, contains, etc."""
 import pytest
 
 from pandas import (
@@ -66,54 +65,3 @@ class TestOverlaps:
         msg = f"`other` must be an Interval, got {type(other).__name__}"
         with pytest.raises(TypeError, match=msg):
             interval.overlaps(other)
-
-
-class TestContains:
-    def test_contains_interval(self, inclusive_endpoints_fixture):
-        interval1 = Interval(0, 1, "both")
-        interval2 = Interval(0, 1, inclusive_endpoints_fixture)
-        assert interval1 in interval1
-        assert interval2 in interval2
-        assert interval2 in interval1
-        assert interval1 not in interval2 or inclusive_endpoints_fixture == "both"
-
-    def test_contains_infinite_length(self):
-        interval1 = Interval(0, 1, "both")
-        interval2 = Interval(float("-inf"), float("inf"), "neither")
-        assert interval1 in interval2
-        assert interval2 not in interval1
-
-    def test_contains_zero_length(self):
-        interval1 = Interval(0, 1, "both")
-        interval2 = Interval(-1, -1, "both")
-        interval3 = Interval(0.5, 0.5, "both")
-        assert interval2 not in interval1
-        assert interval3 in interval1
-        assert interval2 not in interval3 and interval3 not in interval2
-        assert interval1 not in interval2 and interval1 not in interval3
-
-    @pytest.mark.parametrize(
-        "type1",
-        [
-            (0, 1),
-            (Timestamp(2000, 1, 1, 0), Timestamp(2000, 1, 1, 1)),
-            (Timedelta("0h"), Timedelta("1h")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "type2",
-        [
-            (0, 1),
-            (Timestamp(2000, 1, 1, 0), Timestamp(2000, 1, 1, 1)),
-            (Timedelta("0h"), Timedelta("1h")),
-        ],
-    )
-    def test_contains_mixed_types(self, type1, type2):
-        interval1 = Interval(*type1)
-        interval2 = Interval(*type2)
-        if type1 == type2:
-            assert interval1 in interval2
-        else:
-            msg = "^'<=' not supported between instances of"
-            with pytest.raises(TypeError, match=msg):
-                interval1 in interval2
